[files] rename protocol and plugin manifests files
