+++ lib/api/core/entrypoints/index.js
@@ -37,7 +37,7 @@ const
   winston = require('winston'),
   WinstonElasticsearch = require('winston-elasticsearch'),
   WinstonSyslog = require('winston-syslog'),
-  Manifest = require('./manifest'),
+  Manifest = require('./protocolManifest'),
   ClientConnection = require('./clientConnection'),
   removeErrorStack = require('./removeErrorStack'),
   errorsManager = require('../../../config/error-codes/throw')
