+++ lib/api/core/entrypoints/protocolManifest.js
