+++ lib/api/core/plugins/pluginManifest.js
