+++ lib/api/core/plugins/pluginsManager.js
@@ -34,7 +34,7 @@ const
   _ = require('lodash'),
   fs = require('fs'),
   { errors: { KuzzleError } } = require('kuzzle-common-objects'),
-  Manifest = require('./manifest');
+  Manifest = require('./pluginManifest');
 
 /**
  * @class PluginsManager
